test: migrate app_dsl_service tests to testcontainers
